feat: create migration for Location and update Farmer
- Create locations table with coordinate and address fields
- Add location_id foreign key to farmers table
- Replace previous farmers migration with complete version
- Include automatic timestamps (created_at, updated_at)
- Configure appropriate foreign key constraints
